// x-pack/platform/plugins/shared/fleet/public/components/cloud_connector/types.d.ts
interface BaseCloudConnectorCredentials {
    cloudConnectorId?: string;
    name?: string;
}
export interface AwsCloudConnectorCredentials extends BaseCloudConnectorCredentials {
    roleArn?: string;
    externalId?: string | CloudConnectorSecretReference;
}
export interface AzureCloudConnectorCredentials extends BaseCloudConnectorCredentials {
    tenantId?: string | CloudConnectorSecretReference;
    clientId?: string | CloudConnectorSecretReference;
    azure_credentials_cloud_connector_id?: string;
}
export interface GcpCloudConnectorCredentials extends BaseCloudConnectorCredentials {
    serviceAccount?: string;
    audience?: string;
    gcp_credentials_cloud_connector_id?: string | CloudConnectorSecretReference;
}
export type CloudConnectorCredentials = AwsCloudConnectorCredentials | AzureCloudConnectorCredentials | GcpCloudConnectorCredentials;
// ...
export type CloudSetupForCloudConnector = Pick<CloudSetup, 'isCloudEnabled' | 'cloudId' | 'cloudHost' | 'deploymentUrl' | 'serverless' | 'isServerlessEnabled'>;
export interface GetCloudConnectorRemoteRoleTemplateParams {
    cloud: CloudSetupForCloudConnector;
    accountType: AccountType;
    /** IaC template URL to use for generating the cloud connector remote role template. */
    iacTemplateUrl?: string;
}